add hypercube api
